pyproject.toml
[tool.poetry]
name = "e_cards"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.8.3"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
psycopg2 = "^2.8.5"
django-heroku = "^0.3.1"
gunicorn = "^20.0.4"
djoser = "^2.0.3"
django-rest-framework = "^0.1.0"
django-filter = "^2.3.0"
django-cors-headers = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
